Technically finished now, Notes in description
Made so after reach roll it'lkl check for duplicates, essentially if the username is found in the text file, then the new score if higher will replace the score in that line. otherwise itll skip or if its a new entry put it in. The client will then call a response from express/server which will parse through and add new entries to the html displaying the leaderboard.